Rename optimizer model to input model
